Moved SIMD code into arch/avx2.c and arch/sse2.c
The Makefile treats these files specially: avx2.c compiles with -mavx2
and sse2.c with -msse2. This solves two problems.

First problem: The SSE2 code was actually compiling with AVX
instructions because of '-mavx2'. This would cause a crash on hardware
that does not support AVX or AVX2 such as an Atom CPU. (We missed this
testing on the Sandy Bridge lab servers because they do support AVX
even though they do not support AVX2.) Caught by Justin Cormack:
https://groups.google.com/d/msg/snabb-devel/I_3dJvfl7gY/YHgHVzjMq0kJ

Second problem: The Makefile had overlapping rules for building the
SIMD code and this causes warnings to be printed. That is because the
default build rule applied to these files but then we also had
specific rules. This is solved now because the arch/ directory is
excluded from the default rules.
